[UWP][UnitTests] Move sample resource resolver into shared test library
