+++ source/uwp/UWPTestLibrary/UWPTestLibrary.csproj
@@ -116,6 +116,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="RenderedTestResult.cs" />
     <Compile Include="RenderTestHelpers.cs" />
+    <Compile Include="SampleResourceResolver.cs" />
     <Compile Include="StoredTestResultInfo.cs" />
     <Compile Include="TestResultViewModel.cs" />
     <EmbeddedResource Include="Properties\TestLibrary.rd.xml" />
